dense & softmax

+++ src/NeuralNetwork.h
@@ -289,7 +289,7 @@ static float trainNetworkGPU(
 	// printTensor( out_data );
 	// printf("train data end2   ---\n");
 
-	for( int i = 0; i < layers.size(); ++i ){
+	for( int i = 0; i < (int)(layers.size()); ++i ){
 		if( i == 0 ){
 			forwardGPU( layers[i], gpu_in_array );
 		}else{
